requirements.txt
pandas
PySide6
openpyxl
xlrd
python-docx
pyinstaller